+++ braintree/error_codes.py
@@ -29,6 +29,7 @@ class ErrorCodes(object):
         RegionIsTooLong = "81810"
         StreetAddressIsRequired = "81811"
         StreetAddressIsTooLong = "81812"
+        TooManyAddressesPerCustomer = "91818"
 
     class CreditCard(object):
         BillingAddressConflict = "91701"
@@ -80,6 +81,10 @@ class ErrorCodes(object):
         WebsiteIsInvalid = "81616"
         WebsiteIsTooLong = "81615"
 
+    class Descriptor(object):
+        NameFormatIsInvalid = "92201"
+        PhoneFormatIsInvalid = "92202"
+
     class Subscription(object):
         BillingDayOfMonthCannotBeUpdated = "91918"
         BillingDayOfMonthIsInvalid = "91914"
@@ -103,9 +108,11 @@ class ErrorCodes(object):
         PaymentMethodTokenCardTypeIsNotAccepted = "91902"
         PaymentMethodTokenIsInvalid = "91903"
         PaymentMethodTokenNotAssociatedWithCustomer = "91905"
+        PlanBillingFrequencyCannotBeUpdated = "91922"
         PlanIdIsInvalid = "91904"
         PriceCannotBeBlank = "81903"
         PriceFormatIsInvalid = "81904"
+        PriceIsTooLarge = "81923"
         StatusIsCanceled = "81905"
         TokenFormatIsInvalid = "81906"
         TrialDurationFormatIsInvalid = "81907"
@@ -115,6 +122,7 @@ class ErrorCodes(object):
         class Modification(object):
             AmountCannotBeBlank = "92003"
             AmountIsInvalid = "92002"
+            AmountIsTooLarge = "92023"
             CannotEditModificationsOnPastDueSubscription = "92022"
             CannotUpdateAndRemove = "92015"
             ExistingIdIsIncorrectKind = "92020"
@@ -134,26 +142,27 @@ class ErrorCodes(object):
 
     class Transaction(object):
         AmountCannotBeNegative = "81501"
-        AmountIsRequired = "81502"
         AmountIsInvalid = "81503"
+        AmountIsRequired = "81502"
         AmountIsTooLarge = "81528"
         AmountMustBeGreaterThanZero = "81531"
         BillingAddressConflict = "91530"
         CannotBeVoided = "91504"
         CannotRefundCredit = "91505"
         CannotRefundUnlessSettled = "91506"
+        CannotRefundWithSuspendedMerchantAccount = "91538"
         CannotSubmitForSettlement = "91507"
         CreditCardIsRequired = "91508"
-        CustomerDefaultPaymentMethodCardTypeIsNotAccepted = "81509"
         CustomFieldIsInvalid = "91526"
         CustomFieldIsTooLong = "81527"
-        CustomerIdIsInvalid = "91510"
+        CustomerDefaultPaymentMethodCardTypeIsNotAccepted = "81509"
         CustomerDoesNotHaveCreditCard = "91511"
+        CustomerIdIsInvalid = "91510"
         HasAlreadyBeenRefunded = "91512"
-        MerchantAccountNameIsInvalid = "91513" # Deprecated
         MerchantAccountIdIsInvalid = "91513"
         MerchantAccountIsSusped = "91514" # Deprecated
         MerchantAccountIsSuspended = "91514"
+        MerchantAccountNameIsInvalid = "91513" # Deprecated
         OrderIdIsTooLong = "91501"
         PaymentMethodConflict = "91515"
         PaymentMethodDoesNotBelongToCustomer = "91516"
@@ -162,14 +171,17 @@ class ErrorCodes(object):
         PaymentMethodTokenIsInvalid = "91518"
         ProcessorAuthorizationCodeCannotBeSet = "91519"
         ProcessorAuthorizationCodeIsInvalid = "81520"
+        PurchaseOrderNumberIsTooLong = "91537"
         RefundAmountIsTooLarge = "91521"
         SettlementAmountIsTooLarge = "91522"
         SubscriptionDoesNotBelongToCustomer = "91529"
         SubscriptionIdIsInvalid = "91528"
         SubscriptionStatusMustBePastDue = "91531"
+        TaxAmountCannotBeNegative = "81534"
+        TaxAmountFormatIsInvalid = "81535"
+        TaxAmountIsTooLarge = "81536"
         TypeIsInvalid = "91523"
         TypeIsRequired = "91524"
 
         class Options(object):
             VaultIsDisabled = "91525"
-