+++ braintree/transaction.py
@@ -12,9 +12,11 @@ from braintree.address import Address
 from braintree.configuration import Configuration
 from braintree.credit_card import CreditCard
 from braintree.customer import Customer
+from braintree.subscription_details import SubscriptionDetails
 from braintree.resource_collection import ResourceCollection
 from braintree.transparent_redirect import TransparentRedirect
 from braintree.exceptions.not_found_error import NotFoundError
+from braintree.descriptor import Descriptor
 
 class Transaction(Resource):
     """
@@ -114,6 +116,7 @@ class Transaction(Resource):
         * braintree.Transaction.Status.Void
         """
 
+        AuthorizationExpired   = "authorization_expired"
         Authorized             = "authorized"
         Authorizing            = "authorizing"
         Failed                 = "failed"
@@ -305,7 +308,7 @@ class Transaction(Resource):
     @staticmethod
     def create_signature():
         return [
-            "amount", "customer_id", "merchant_account_id", "order_id", "payment_method_token", "type",
+            "amount", "customer_id", "merchant_account_id", "order_id", "payment_method_token", "purchase_order_number", "shipping_address_id", "tax_amount", "tax_exempt", "type",
             {
                 "credit_card": [
                     "token", "cardholder_name", "cvv", "expiration_date", "expiration_month", "expiration_year", "number"
@@ -332,11 +335,12 @@ class Transaction(Resource):
             },
             {
                 "options": [
-                    "store_in_vault", "submit_for_settlement", "add_billing_address_to_payment_method",
-                    "store_shipping_address_in_vault"
+                    "store_in_vault", "store_in_vault_on_success", "submit_for_settlement",
+                    "add_billing_address_to_payment_method", "store_shipping_address_in_vault"
                 ]
             },
-            {"custom_fields": ["__any_key__"]}
+            {"custom_fields": ["__any_key__"]},
+            {"descriptor": ["name", "phone"]}
         ]
 
     def __init__(self, gateway, attributes):
@@ -349,6 +353,8 @@ class Transaction(Resource):
         Resource.__init__(self, gateway, attributes)
 
         self.amount = Decimal(self.amount)
+        if self.tax_amount:
+            self.tax_amount = Decimal(self.tax_amount)
         if "billing" in attributes:
             self.billing_details = Address(gateway, attributes.pop("billing"))
         if "credit_card" in attributes:
@@ -363,6 +369,10 @@ class Transaction(Resource):
             self.discounts = [Discount(gateway, discount) for discount in self.discounts]
         if "status_history" in attributes:
             self.status_history = [StatusEvent(gateway, status_event) for status_event in self.status_history]
+        if "subscription" in attributes:
+            self.subscription_details = SubscriptionDetails(attributes.pop("subscription"))
+        if "descriptor" in attributes:
+            self.descriptor = Descriptor(gateway, attributes.pop("descriptor"))
 
     @property
     def refund_id(self):
@@ -382,7 +392,6 @@ class Transaction(Resource):
         """
         The vault credit card associated with this transaction
         """
-
         if self.credit_card_details.token is None:
             return None
         return self.gateway.credit_card.find(self.credit_card_details.token)
@@ -392,5 +401,6 @@ class Transaction(Resource):
         """
         The vault customer associated with this transaction
         """
-
+        if self.customer_details.id is None:
+            return None
         return self.gateway.customer.find(self.customer_details.id)