+++ tests/integration/test_subscription.py
@@ -138,6 +138,15 @@ class TestSubscription(unittest.TestCase):
         self.assertEquals("sale", transaction.type)
         self.assertEquals(subscription.id, transaction.subscription_id)
 
+    def test_create_has_transaction_with_billing_period_dates(self):
+        subscription = Subscription.create({
+            "payment_method_token": self.credit_card.token,
+            "plan_id": TestHelper.trialless_plan["id"],
+        }).subscription
+        transaction = subscription.transactions[0]
+        self.assertEquals(subscription.billing_period_start_date, transaction.subscription_details.billing_period_start_date)
+        self.assertEquals(subscription.billing_period_end_date, transaction.subscription_details.billing_period_end_date)
+
     def test_create_returns_a_transaction_if_transaction_is_declined(self):
         result = Subscription.create({
             "payment_method_token": self.credit_card.token,
@@ -418,6 +427,48 @@ class TestSubscription(unittest.TestCase):
         self.assertEquals(
             ErrorCodes.Subscription.Modification.QuantityIsInvalid,
             result.errors.for_object("subscription").for_object("add_ons").for_object("update").for_index(1).on("quantity")[0].code
+        )
+
+    def test_descriptors_accepts_name_and_phone(self):
+        result = Subscription.create({
+            "payment_method_token": self.credit_card.token,
+            "plan_id": TestHelper.trialless_plan["id"],
+            "descriptor": {
+                "name": "123*123456789012345678",
+                "phone": "3334445555"
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        subscription = result.subscription
+        self.assertEquals("123*123456789012345678", subscription.descriptor.name)
+        self.assertEquals("3334445555", subscription.descriptor.phone)
+
+        transaction = subscription.transactions[0]
+        self.assertEquals("123*123456789012345678", transaction.descriptor.name)
+        self.assertEquals("3334445555", transaction.descriptor.phone)
+
+    def test_descriptors_has_validation_errors_if_format_is_invalid(self):
+        result = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "credit_card": {
+                "number": "4111111111111111",
+                "expiration_date": "05/2009"
+            },
+            "descriptor": {
+                "name": "badcompanyname12*badproduct12",
+                "phone": "%bad4445555"
+            }
+        })
+        self.assertFalse(result.is_success)
+        transaction = result.transaction
+        self.assertEquals(
+            ErrorCodes.Descriptor.NameFormatIsInvalid,
+            result.errors.for_object("transaction").for_object("descriptor").on("name")[0].code
+        )
+        self.assertEquals(
+            ErrorCodes.Descriptor.PhoneFormatIsInvalid,
+            result.errors.for_object("transaction").for_object("descriptor").on("phone")[0].code
         )
 
     def test_find_with_valid_id(self):
@@ -751,6 +802,28 @@ class TestSubscription(unittest.TestCase):
         self.assertEquals(None, subscription.discounts[0].number_of_billing_cycles)
         self.assertTrue(subscription.discounts[0].never_expires)
 
+    def test_update_descriptor_name_and_phone(self):
+        result = Subscription.create({
+            "payment_method_token": self.credit_card.token,
+            "plan_id": TestHelper.trialless_plan["id"],
+            "descriptor": {
+                "name": "123*123456789012345678",
+                "phone": "1234567890"
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        subscription = result.subscription
+        updated_subscription = Subscription.update(subscription.id, {
+            "descriptor": {
+                "name": "999*99",
+                "phone": "1234567890"
+            }
+        }).subscription
+
+        self.assertEquals("999*99", updated_subscription.descriptor.name)
+        self.assertEquals("1234567890", updated_subscription.descriptor.phone)
+
     def test_cancel_with_successful_response(self):
         subscription = Subscription.create({
             "payment_method_token": self.credit_card.token,
@@ -942,6 +1015,24 @@ class TestSubscription(unittest.TestCase):
         self.assertTrue(TestHelper.includes(collection, subscription_1000))
         self.assertFalse(TestHelper.includes(collection, subscription_900))
 
+    def test_search_on_transaction_id(self):
+        subscription_found = Subscription.create({
+            "payment_method_token": self.credit_card.token,
+            "plan_id": TestHelper.trialless_plan["id"],
+        }).subscription
+
+        subscription_not_found = Subscription.create({
+            "payment_method_token": self.credit_card.token,
+            "plan_id": TestHelper.trialless_plan["id"],
+        }).subscription
+
+        collection = Subscription.search(
+            SubscriptionSearch.transaction_id == subscription_found.transactions[0].id
+        )
+
+        self.assertTrue(TestHelper.includes(collection, subscription_found))
+        self.assertFalse(TestHelper.includes(collection, subscription_not_found))
+
     def test_search_on_id(self):
         subscription_found = Subscription.create({
             "id": "find_me_%s" % random.randint(1,1000000),
@@ -958,6 +1049,26 @@ class TestSubscription(unittest.TestCase):
         collection = Subscription.search([
             SubscriptionSearch.id.starts_with("find_me")
         ])
+
+        self.assertTrue(TestHelper.includes(collection, subscription_found))
+        self.assertFalse(TestHelper.includes(collection, subscription_not_found))
+
+    def test_search_on_next_billing_date(self):
+        subscription_found = Subscription.create({
+            "payment_method_token": self.credit_card.token,
+            "plan_id": TestHelper.trialless_plan["id"]
+        }).subscription
+
+        subscription_not_found = Subscription.create({
+            "payment_method_token": self.credit_card.token,
+            "plan_id": TestHelper.trial_plan["id"]
+        }).subscription
+
+        next_billing_date_cutoff = datetime.today() + timedelta(days=5)
+
+        collection = Subscription.search(
+            SubscriptionSearch.next_billing_date >= next_billing_date_cutoff
+        )
 
         self.assertTrue(TestHelper.includes(collection, subscription_found))
         self.assertFalse(TestHelper.includes(collection, subscription_not_found))